pc: checkpoint LP64 bring-up fixes

+++ include/libforest/emu64/emu64.hpp
@@ -639,7 +639,7 @@ class emu64 : public emu64_print {
     void print_combine(u64 combine);
     void print_combine_tev(u64 combine_tev);
     void print_guMtxXFM1F_dol2(MtxP mtx, GXProjectionType type, float x, float y, float z);
-    u32 seg2k0(u32 seg);
+    uintptr_t seg2k0(u32 seg);
     void setup_texture_tile(int tile);
     void setup_1tri_2tri_1quad(unsigned int vtx_idx);
     void draw_1tri_2tri_1quad(unsigned int n_verts, ...);
@@ -747,7 +747,7 @@ private:
     /* 0x0060 */ ucode_info* ucode_info_p;
     /* 0x0064 */ int ucode_type; // maybe?
     /* 0x0068 */ int _0068;      /* ??? */
-    /* 0x006C */ u32 segments[EMU64_NUM_SEGMENTS];
+    /* 0x006C */ uintptr_t segments[EMU64_NUM_SEGMENTS];
     /* 0x00AC */ uintptr_t DL_stack[DL_MAX_STACK_LEVEL];
     /* 0x00F4 */ s8 DL_stack_level;
     /* 0x00F8 */ u32 othermode_high;
